week3teht/src/main/java/tel/kontra/Account.java
package tel.kontra;

public class Account {
    
    double balance;

    public Account() {
        this.balance = 0.0;
    }

    public void deposit(double amount) {
        if (amount > 0) {
            this.balance += amount;
        }
    }

    public double withdraw(double amount) {
        if(amount > 0 && amount <= this.balance) {
            this.balance -= amount;
            return amount;
        } else {
            return 0.0;
        }
    }

    public double getBalance() {
        return this.balance;
    }

}